refactor: restructure frontend for cleaner code and better maintainability
