+++ frontend/src/indexed-documents/styles/indexed-documents.css
@@ -1,0 +1,199 @@
+.indexed-documents {
+    background: white;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    padding: 24px;
+    margin-bottom: 30px;
+}
+
+.indexed-documents h2 {
+    color: #2c5282;
+    margin-bottom: 20px;
+}
+
+.section-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 20px;
+}
+
+.toggle-button {
+    background-color: #e2e8f0;
+    color: #4a5568;
+    border: none;
+    padding: 6px 12px;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 12px;
+    transition: background-color 0.3s ease;
+}
+
+.toggle-button:hover {
+    background-color: #cbd5e0;
+}
+
+.status-summary {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    gap: 16px;
+    margin-bottom: 24px;
+}
+
+.status-card {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+    background-color: #f7fafc;
+    padding: 16px;
+    border-radius: 6px;
+    border: 1px solid #e2e8f0;
+}
+
+.status-icon {
+    font-size: 24px;
+    flex-shrink: 0;
+}
+
+.status-info {
+    display: flex;
+    flex-direction: column;
+}
+
+.status-number {
+    font-size: 20px;
+    font-weight: bold;
+    color: #2d3748;
+    line-height: 1;
+}
+
+.status-label {
+    font-size: 12px;
+    color: #718096;
+    margin-top: 2px;
+}
+
+.status-badge {
+    font-size: 12px;
+    font-weight: 600;
+    padding: 2px 6px;
+    border-radius: 12px;
+    line-height: 1;
+}
+
+.status-badge.success {
+    background-color: #c6f6d5;
+    color: #276749;
+}
+
+.status-badge.error {
+    background-color: #fed7d7;
+    color: #c53030;
+}
+
+.loading-state,
+.empty-state {
+    text-align: center;
+    padding: 40px 20px;
+    color: #666;
+}
+
+.loading-state {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 12px;
+}
+
+.empty-state p {
+    margin: 8px 0;
+}
+
+.empty-state p:first-child {
+    font-size: 18px;
+    margin-bottom: 8px;
+}
+
+.documents-list {
+    border-top: 1px solid #e2e8f0;
+    padding-top: 20px;
+}
+
+.list-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 16px;
+}
+
+.list-header h3 {
+    color: #4a5568;
+    font-size: 16px;
+    margin: 0;
+}
+
+.refresh-mini-button {
+    background-color: #4299e1;
+    color: white;
+    border: none;
+    padding: 4px 8px;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 12px;
+    transition: background-color 0.3s ease;
+}
+
+.refresh-mini-button:hover {
+    background-color: #3182ce;
+}
+
+.documents-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+    gap: 12px;
+}
+
+.document-card {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+    background-color: #f8f9fa;
+    border: 1px solid #e9ecef;
+    border-radius: 6px;
+    padding: 12px;
+    transition: all 0.3s ease;
+}
+
+.document-card:hover {
+    border-color: #4299e1;
+    box-shadow: 0 2px 8px rgba(66, 153, 225, 0.15);
+}
+
+.document-icon {
+    font-size: 20px;
+    flex-shrink: 0;
+}
+
+.document-info {
+    display: flex;
+    flex-direction: column;
+    min-width: 0;
+}
+
+.document-name {
+    font-weight: 500;
+    color: #2d3748;
+    font-size: 14px;
+    word-break: break-all;
+    line-height: 1.3;
+}
+
+.document-type {
+    font-size: 11px;
+    color: #718096;
+    background-color: #e2e8f0;
+    padding: 2px 6px;
+    border-radius: 8px;
+    margin-top: 4px;
+    width: fit-content;
+}
