+++ frontend/src/qa/interfaces/qa.interfaces.ts
@@ -1,0 +1,10 @@
+export interface Citation {
+    text: string;
+    document_name: string;
+}
+
+export interface AskResponse {
+    question: string;
+    answer: string;
+    citations: Citation[];
+}
