+++ frontend/src/search/interfaces/search.interfaces.ts
@@ -1,0 +1,11 @@
+export interface SearchResult {
+    text: string;
+    document_name: string;
+    relevance_score: number;
+}
+
+export interface SearchResponse {
+    query: string;
+    results: SearchResult[];
+    total_results: number;
+}
