+++ frontend/src/uploader/styles/uploader.css
@@ -1,0 +1,191 @@
+.file-uploader {
+    background: white;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    padding: 24px;
+}
+
+.uploader-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 20px;
+}
+
+.file-uploader h2 {
+    color: #2c5282;
+    margin: 0;
+}
+
+.test-connection-button {
+    background-color: #4299e1;
+    color: white;
+    border: none;
+    padding: 8px 16px;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 14px;
+    transition: background-color 0.3s ease;
+}
+
+.test-connection-button:hover:not(:disabled) {
+    background-color: #3182ce;
+}
+
+.test-connection-button:disabled {
+    background-color: #a0aec0;
+    cursor: not-allowed;
+}
+
+.drop-zone {
+    border: 2px dashed #cbd5e0;
+    border-radius: 8px;
+    padding: 40px;
+    text-align: center;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    background-color: #f8f9fa;
+}
+
+.drop-zone:hover,
+.drop-zone.drag-active {
+    border-color: #4299e1;
+    background-color: #ebf8ff;
+}
+
+.upload-icon {
+    font-size: 48px;
+    margin-bottom: 16px;
+}
+
+.drop-zone h3 {
+    color: #2d3748;
+    margin-bottom: 8px;
+}
+
+.drop-zone p {
+    color: #666;
+}
+
+.files-ready {
+    text-align: center;
+}
+
+.selected-files-list {
+    margin: 16px 0;
+    display: grid;
+    gap: 8px;
+    max-height: 200px;
+    overflow-y: auto;
+    padding: 12px;
+    background-color: #f8f9fa;
+    border-radius: 6px;
+    border: 1px solid #e9ecef;
+}
+
+.file-item {
+    padding: 8px 12px;
+    background-color: #e6fffa;
+    border-radius: 4px;
+    text-align: left;
+}
+
+.process-actions {
+    display: flex;
+    gap: 12px;
+    justify-content: center;
+    margin: 20px 0;
+}
+
+.process-button {
+    background-color: #38a169;
+    color: white;
+    border: none;
+    padding: 12px 24px;
+    border-radius: 6px;
+    cursor: pointer;
+    font-size: 16px;
+    font-weight: 600;
+    transition: background-color 0.3s ease;
+    min-width: 180px;
+}
+
+.process-button:hover:not(:disabled) {
+    background-color: #2f855a;
+}
+
+.process-button:disabled {
+    background-color: #a0aec0;
+    cursor: not-allowed;
+}
+
+.cancel-button {
+    background-color: #718096;
+    color: white;
+    border: none;
+    padding: 12px 24px;
+    border-radius: 6px;
+    cursor: pointer;
+    font-size: 16px;
+    transition: background-color 0.3s ease;
+}
+
+.cancel-button:hover:not(:disabled) {
+    background-color: #4a5568;
+}
+
+.cancel-button:disabled {
+    background-color: #a0aec0;
+    cursor: not-allowed;
+}
+
+.upload-status {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 12px;
+    padding: 20px;
+}
+
+.spinner {
+    width: 24px;
+    height: 24px;
+    border: 3px solid #e2e8f0;
+    border-top: 3px solid #4299e1;
+    border-radius: 50%;
+    animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+    0% {
+        transform: rotate(0deg);
+    }
+
+    100% {
+        transform: rotate(360deg);
+    }
+}
+
+.indexed-documents-summary {
+    text-align: center;
+}
+
+.file-list {
+    margin: 20px 0;
+    display: grid;
+    gap: 8px;
+}
+
+.clear-button {
+    background-color: #e53e3e;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+.clear-button:hover {
+    background-color: #c53030;
+}
